+++ src/main/java/com/game/bingo/Repository/NumberRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NumberRepository extends CrudRepository<Number, Integer> {
+public interface NumberRepository extends CrudRepository<Number, Long> {
 }
