+++ src/main/java/com/game/bingo/Repository/PlayerRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PlayerRepository extends CrudRepository<Player, Integer> {
+public interface PlayerRepository extends CrudRepository<Player, Long> {
     List<Player> findByRoom(Room room);
 }
